refactor: rename folder from api to server
