+++ web-insights-server/build.gradle.kts
