+++ web-insights-server/src/main/java/io/web_insights/backend/SpringBootWebInsightsApplication.java
@@ -1,4 +1,4 @@
-package org.webinsights.api;
+package io.web_insights.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
