+++ web-insights-server/src/main/java/io/web_insights/backend/config/JwtAuthConverter.java
@@ -1,4 +1,4 @@
-package org.webinsights.api.config;
+package io.web_insights.backend.config;
 
 import java.util.Collection;
 import java.util.Map;
