+++ web-insights-server/src/main/java/io/web_insights/backend/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package org.webinsights.api.config;
+package io.web_insights.backend.config;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
