+++ web-insights-server/src/main/java/io/web_insights/backend/workflow/WorkflowService.java
@@ -1,4 +1,4 @@
-package org.webinsights.api.workflow;
+package io.web_insights.backend.workflow;
 
 import jakarta.persistence.EntityNotFoundException;
 import java.util.ArrayList;
