+++ web-insights-server/src/main/resources/application.properties
