+++ web-insights-server/src/test/java/io/web_insights/backend/SpringBootWebInsightsApplicationTests.java
@@ -1,4 +1,4 @@
-package org.webinsights.api;
+package io.web_insights.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
